OOP-13 super

+++ src/corporation/ShoeCard.kt
@@ -4,7 +4,7 @@ class ShoeCard(
     name: String,
     brand: String,
     price: Int,
-    val size: Double,
+    val size: Float,
 ) : ProductCard(name, brand, price) {
 
     override fun printInfo() {
